data_utils/requirements.txt
numpy==1.17.4
matplotlib==3.1.2
openpyxl==3.0.3
opencv_python==4.8.1.78
pandas==0.25.3
matplotlib_venn==0.11.6
scipy==1.3.3
tqdm==4.65.0
fire==0.2.1
plotly==5.1.0
tabulate==0.8.9
Pillow==10.2.0
PyYAML==6.0.1
skimage==0.0
